requirements.txt
pymupdf
spacy
transformers
sentence-transformers
faiss-cpu
numpy
chainlit
langchain
langchain-openai
langchain-core
arize-phoenix                             
arize-phoenix-evals                       
arize-phoenix-otel                        